+++ server/CHANGELOG.md
@@ -6,3 +6,4 @@
 - Dotenv validation
 - Bare minimum express with CORS is setup
 - Working on the basic structure of the database in prisma/schema.prisma
+- Created envVars.ts to handle validation of environment variables removing some boilerplate code from main.ts